==> notebook_example/02.word_embedding_train.ipynb <==
import logging
import gensim
import pandas as pd
import sys
import os
import multiprocessing
from ast import literal_eval
from collections import namedtuple
from gensim.models.doc2vec import Doc2Vec, TaggedDocument
from gensim.test.utils import common_texts, get_tmpfile
from sklearn.model_selection import train_test_split
import json
sys.path.append(os.pardir)
from models.word_embedding import word2vec_model, doc2vec_model

## word embedding training

model hyperparameter change using kwargs
+ function : choose doc2vec or word2vec
+ min_cound : word embedding 

you can check more detail code in ./models/word_embedding.py

def main(**kwargs):
    data = pd.read_excel("../assets/data/doc_set_final_version3.xlsx")
    data.token = data.token.apply(lambda x : literal_eval(x))
    data = data.sample(frac=1, random_state=1234)

    token_list = data.token.tolist()
    target = data[['new_class', 'new_small_class']]
    train_x_data, test_x_data, train_y, test_y = train_test_split(token_list, target,
                                                                    test_size=0.3,
                                                                    stratify=target,
                                                                    shuffle=True,
                                                                    random_state=1234)

    model_select = kwargs['function']
    if model_select == 'w2v':
        print("모델 학습")
        word2vec_kargs = {'num_features':kwargs['num_features'],
                          'num_workers':4,
                          'window':kwargs['window'],
                          'seed':1234,
                          'min_word_count':kwargs['min_count'],
                          'min_alpha':kwargs['min_alpha'],
                          'iter': kwargs['iter']}
        model = word2vec_model(train_x_data, **word2vec_kargs)
        print("모델 저장")
        model_name = '../model_save/embedding_model/Word2vec1.model'
        model.save(model_name)

    elif model_select == 'd2v':
        TaggedDocument = namedtuple('TaggedDocument', 'words tags')
        tagged_train_docs = [TaggedDocument(d, [c[1]['new_class'], c[1]['new_small_class']]) for d, c in zip(train_x_data, train_y.iterrows())]
        print("모델 학습")
        doc2vec_kargs = {'size':kwargs['num_features'], #300
                         'window':kwargs['window'], #10
                         'min_count':kwargs['min_count'], # 3
                         'alpha':0.025,
                         'min_alpha':kwargs['min_alpha'], # 0.025
                         'workers':4,
                         'seed':1234,
                         'iter':kwargs['iter']} # 30
        model = doc2vec_model(tagged_train_docs, **doc2vec_kargs)
        print("모델 저장")
        model.save('../model_save/embedding_model/Doc2vec_new.model')

    else:
        print("2가지 방식 중에 고르시오")

if __name__ == '__main__':
#     args = parse_args()
#     print('Called with args:')
#     kwargs = dict(args._get_kwargs())
    kwargs = {'function': 'w2v', 'iter': 30, 'min_alpha': 0.025, 'min_count': 5, 'num_features': 300, 'window': 8}
    print(kwargs)
    main(**kwargs)
