==> notebook_example/07.evaluation.ipynb <==
import os
import sys
import json
import matplotlib.pyplot as plt
import numpy as np
import tensorflow as tf
import tensorflow_addons as tfa
sys.path.append(os.pardir)
from gensim.models.doc2vec import Doc2Vec, TaggedDocument
from gensim.models.word2vec import Word2Vec
from models.word_embedding import get_embedding_matrix
from models.cnn_kr import CNNClassifier

w2v_model_name = '../model_save/embedding_model/Word2vec1.model'
word_vectorizer = Word2Vec.load(w2v_model_name)

pre_trained_name = '../model_save/embedding_model/trained_word2vec1.model'
pre_trained_w2v = Word2Vec.load(pre_trained_name)

## 01. Data Load

DATA_IN_PATH = '../assets/data/npy_data/2020-05-31/'

# Data save label
TRAIN_INPUT_DATA = 'train_input.npy'
TEST_INPUT_DATA = 'test_input.npy'
DATA_CONFIGS = 'data_configs.json'
SEQ_CONFIGS = 'seq_configs_bt.json'

# Train label save file name
TRAIN_LABEL_DATA = 'train_label.npy'
TEST_LABEL_DATA = 'test_label.npy'
TRAIN_LABEL_SMALL = 'train_label_small.npy'
TEST_LABEL_SMALL = 'test_label_small.npy'

train_X = np.load(open(DATA_IN_PATH + TRAIN_INPUT_DATA, 'rb'))
train_Y = np.load(open(DATA_IN_PATH + TRAIN_LABEL_DATA, 'rb'))
test_X = np.load(open(DATA_IN_PATH + TEST_INPUT_DATA, 'rb'))
test_Y = np.load(open(DATA_IN_PATH + TEST_LABEL_DATA, 'rb'))
train_YS = tf.one_hot(train_Y, 43)
test_YS = tf.one_hot(test_Y, 43)
data_configs = json.load(open(DATA_IN_PATH + DATA_CONFIGS, 'r'))
vocab_size = data_configs['vocab_size'] + 1
print("vocab_size : ", vocab_size)
word_index = data_configs['vocab']

### best parameter each neural network

cnn_kargs = {'vocab_size': vocab_size,
             'embedding_size': 300, 
             'num_filters': 128, 
             'dropout_rate': 0.5, 
             'hidden_dimension': 500, 
             'train_mode': 'non_static', 
             'output_dimension': 43, 
             'trainable':True,
             'optimizer':'adam',
             'model_name': 'cnn_non_static_adam_w2v'}

rnn_kargs = {'vocab_size': vocab_size,
             'embedding_size': 300, 
             'dropout_rate': 0.5, 
             'lstm_dimension': 128, 
             'dense_dimension': 64, 
             'train_mode': 'pt', 
             'output_dimension': 43, 
             'optimizer':'radam',
             'model_name': 'lstm_pt_radam_pt_w2v'}

## 02. model evaluation

### 02-1. big_class

from models.model_evaluation import model_build, return_two_calss_acc

### cnn_non_static w2v

cnn_model = model_build(train_X, word_index, word_vectorizer, **cnn_kargs)

cnn_model.evaluate(test_X, test_YS)

### predict_proba two class

two_class_acc = return_two_calss_acc(cnn_model, test_X, test_Y)

two_class_acc

### bidirection LSTM

rnn_model = model_build(train_X, word_index, word_vectorizer, **rnn_kargs)

rnn_model.evaluate(test_X, test_YS)

two_class_acc = return_two_calss_acc(rnn_model, test_X, test_Y)

two_class_acc

### AC-BLSTM

### 02-2. small_class

### CNN

test_Y = np.load(open(DATA_IN_PATH + TEST_LABEL_SMALL, 'rb'))
test_YS = tf.one_hot(test_Y, 455)

cnn_kargs['output_dimension'] = 455
rnn_kargs['output_dimension'] = 455

cnn_model_small = model_build(train_X, word_index, word_vectorizer, **cnn_kargs)

cnn_model_small.evaluate(test_X, test_YS)

two_class_acc_small = return_two_calss_acc(cnn_model_small, test_X, test_Y)

two_class_acc_small

### LSTM

rnn_model_small = model_build(train_X, word_index, word_vectorizer, **rnn_kargs)

rnn_model_small.evaluate(test_X, test_YS)

two_class_acc_small = return_two_calss_acc(rnn_model_small, test_X, test_Y)

two_class_acc_small

### 03. application
